报支类别集团设置

+++ jlkj-modules/jlkj-finance/src/main/java/com/jlkj/finance/ap/service/impl/FinanceApItemServiceImpl.java
@@ -1,5 +1,6 @@
 package com.jlkj.finance.ap.service.impl;
 
+import com.jlkj.common.core.exception.ServiceException;
 import com.jlkj.common.core.utils.DateUtils;
 import com.jlkj.common.core.utils.uuid.IdUtils;
 import com.jlkj.common.security.utils.SecurityUtils;
@@ -10,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 
 /**
  * 报支类别集团设置Service业务层处理
@@ -32,7 +34,20 @@ public class FinanceApItemServiceImpl implements IFinanceApItemService
     @Override
     public FinanceApItem selectFinanceApItemById(String id)
     {
+
         return financeApItemMapper.selectFinanceApItemById(id);
+    }
+    /**
+     * 查询报支类别集团设置
+     *
+     * @param id 报支类别集团设置主键
+     * @return 报支类别集团设置
+     */
+    @Override
+    public FinanceApItem selectFinanceApItemByIdcal(String id)
+    {
+        FinanceApItem financeApItem = financeApItemMapper.selectFinanceApItemByIdcal(id);
+        return financeApItem;
     }
 
     /**
@@ -56,6 +71,10 @@ public class FinanceApItemServiceImpl implements IFinanceApItemService
     @Override
     public int insertFinanceApItem(FinanceApItem financeApItem)
     {
+        int count = financeApItemMapper.updateCheckUnique(financeApItem.getItemNo());
+        if(count!=0){
+            throw new ServiceException("报支类别编码："+financeApItem.getItemNo()+"已经存在，不可重复新增！");
+        }
         financeApItem.setId(IdUtils.fastSimpleUUID());
         financeApItem.setCreateTime(DateUtils.getNowDate());
         financeApItem.setCreateBy(SecurityUtils.getUsername());
@@ -87,6 +106,11 @@ public class FinanceApItemServiceImpl implements IFinanceApItemService
     @Override
     public int deleteFinanceApItemByIds(String[] ids)
     {
+        List<Map<String, String>> itemNos = financeApItemMapper.selectItemNosByIds(ids);
+        if(itemNos.size()>0){
+            throw new ServiceException("删除失败，存在细项资料, 请先删除细项资料！");
+        }
+
         return financeApItemMapper.deleteFinanceApItemByIds(ids);
     }
 
@@ -100,5 +124,15 @@ public class FinanceApItemServiceImpl implements IFinanceApItemService
     public int deleteFinanceApItemById(String id)
     {
         return financeApItemMapper.deleteFinanceApItemById(id);
+    }
+
+    /**
+     * 选取集团报支类别(公司级报支类别用)
+     * @param companyId
+     * @return
+     */
+    @Override
+    public List<Map<String,Object>> selectItemAndDetailList(String companyId){
+        return financeApItemMapper.selectItemAndDetailList(companyId);
     }
 }
